renamed all to gss and remove requirements.txt file
