+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include src/AGI/fonts/*.ttf
+include src/GSS/fonts/*.ttf
