+++ src/GSS/io/format.py
@@ -1,5 +1,5 @@
 ##############################################################
-# Project: Alps GPU Insight
+# Project: GPU saturation scorer
 #
 # File Name: format.py
 #
