+++ src/GSS/io/json_io.py
@@ -1,46 +1,48 @@
-##############################################################
-# Project: Alps GPU Insight
+###############################################################
+# Project: GPU saturation scorer
 #
-# File Name: binary_io.py
+# File Name: json_io.py
 #
 # Description:
-# This file contains the class to handle binary data input/output.
+# This file contains the JSONDataIO class, which is used to
+# handle JSON data input/output.
 #
 # Authors:
 # Marcel Ferrari (CSCS)
 #
 ###############################################################
 
-import pickle
+import json
 import os
-from AGI.io.base_io import BaseIO
+from GSS.io.base_io import BaseIO
 
-# This class is used to handle JSON data input/output
-class BinaryDataIO(BaseIO):
+class JSONDataIO(BaseIO):
     """
     Description:
-    This class provides a context manager interface to read and write binary data.
-    It inherits from the BaseIO class.
+    This class is used to handle JSON data input/output.
 
     Attributes:
-    - file: The file to read/write data to/from.
-    - force_overwrite: A flag to force overwrite of the file.
+    - file (str): Path to the JSON file.
+    - force_overwrite (bool): Flag to force overwrite of existing file.
 
     Methods:
-    - dump(self, metadata: dict, data: dict): Dump metadata and data to a binary file.
-    - load(self) -> tuple: Load metadata and data from a binary file.
+    - dump(self, metadata: dict, data: dict) -> None: Dump metadata and data to a JSON file.
+    - load(self) -> tuple: Load metadata and data from a JSON file.
 
     Notes:
-    - Under the hood, this class uses the pickle module to serialize and deserialize data.
+    - This class uses the json module to serialize and deserialize data.
+    - JSON is the default format for data output as it is human-readable.
+      If you need to save space, consider using the BinaryDataIO class.
     """
+
     def __init__(self, file: str, force_overwrite: bool = False) -> None:
         """
         Description:
         Constructor method.
 
         Parameters:
-        - file: The file to read/write data to/from.
-        - force_overwrite: A flag to force overwrite of the file.
+        - file (str): Path to the JSON file.
+        - force_overwrite (bool): Flag to force overwrite of existing file.
 
         Returns:
         - None
@@ -55,44 +57,44 @@ class BinaryDataIO(BaseIO):
     def dump(self, metadata: dict, data: dict) -> None:
         """
         Description:
-        This method dumps metadata and data to a binary file.
+        This method dumps metadata and data to a JSON file.
 
         Parameters:
-        - metadata: A dictionary containing metadata.
-        - data: A dictionary containing data.
+        - metadata (dict): A dictionary containing metadata.
+        - data (dict): A dictionary containing data.
 
         Returns:
         - None
 
         Notes:
-        - This method writes metadata and data to a binary file using the pickle module.
         - The data dictionary should contain a mapping from GPU ID to a dictionary containing metric: values pairs.
           This is important as the GPU IDs are NOT stored explicitly in the metadata.
         """
         # Create directory if necessary
         dirname = os.path.dirname(self.file)
+        print(dirname)
         os.makedirs(dirname, exist_ok=True)
 
         # Combine metadata and data and write to file
-        with open(self.file, 'wb+') as f:
-            pickle.dump({'metadata': metadata, 'data': data}, f)
+        with open(self.file, 'w+') as f:
+            json.dump({'metadata': metadata, 'data': data}, f)
 
     # This function loads the data from a JSON file
     def load(self) -> tuple:
         """
         Description:
-        This method loads metadata and data from a binary file.
+        This method loads metadata and data from a JSON file.
 
         Parameters:
         - None
 
         Returns:
-        - A tuple containing metadata and data.
+        - tuple: A tuple containing metadata and data.
+          This might be changed in the future when CPU/MPI/NCCL support is added.
         """
-
         # Read data from file
-        with open(self.file, 'rb') as f:
-            data = pickle.load(f)
+        with open(self.file, 'r') as f:
+            data = json.load(f)
 
         # Try to return metadata and data
         try:
